Rename createChildren
